Add framework to work with a Scene
The basic necessities for rendering a scene
is added. It includes creating the main OpenGL
context (Entity), which is a hidden window, and
some OpenGL resource wrappers: texture, framebuffer.
Scene is now capable to render into a texture,
with a simple color, nothing more yet.

Also a test is added for simple scene rendering.