Add window
Added the Window class which renders the rendered scene.
Vertex buffer, vertex array and quad OpenGL
resources are also added which are needed, because
that's the canvas we draw the rendered scene onto.
To display the quad, we need simple shaders,
and shader compilation capabilities.

Also, tests are added to ensure window
works by itself properly, and that it
displays the rendered scene properly,