Fixed gif creation issues, added gif animation to readme
Gif creation had an issue. This is a problem with FFmpeg
itself, the frame conversion needs an intermediate step.
This intermediate step (`intermediate_yuv420p_conversion`)
is now enabled by default, but has no effect on videos
where the native pixel format is YUV420P.

The Lorenz attractor example has been also modified slightly,
so that now it creates a gif image instead of mp4, and the result
is also uploaded and inserted into the readme file.