Scene, Window and Video creation returns a shared pointer
Previously the Scene, Window, Video create function
returned an Expected type which wrapped a simple Scene,
Window, Video type. This is very limited, because that can only be
moved with move semantics. (This is mandatory to not have references
to the same impl.) Now, a shared pointer is created, which makes it
explicit that there are multiple references to the same object,
and also allows us to have only one impl reference for one object.