Fix video encoding bug, now encodes all the frames in the buffer
There was a bug with video encoding. When the codec returns with
EAGAIN error message, that indicates it needs more frames, and does
not return any packet. The buffer can stay non-empty even after
the last frame was sent. In this case, we should flush the codec,
and get all the remaining packets too.
See <https://ffmpeg.org/doxygen/3.3/group__lavc__encdec.html>.