Huge performance improvement for surface
SurfaceVisual previously needed to convert the SurfaceData
to the OpenGL representation every time a new
SurfaceVisual was created, or `set_surface_data` was called
on the SurfaceVisual. Now, this conversion does not happen,
the SurfaceData holds it's data in it's OpenGL format,
which means that the surfaces can be generad once,
and then rendered without transformation, which makes a huge difference.

Unfortunately this way more memory is needed, but this is not a big
problem, because the old way of handling conversion still is an option
by generated the SurfaceData just before calling `set_surface_data`,
and then releasing the memory right after that.