Memory improvements for surface using shader buffers and vertex indexing
Now the SurfaceData and SurfaceVisual uses vertex indexing.
This means, that the unique vertex information are stored once,
and then there is an indexing which describes which unique information
is stored where for each drawn vertex. For this, shader buffer was
needed, because for the flat mode, not every data is shared at the same
time: color can be different for the same position, but there might be
colors and normals which need to be stored only once for several draw
vertices.